pyproject.toml
[project]
name = "oc-02-modetrends"
version = "0.1.0"
description = ""
authors = [
    {name = "Christophe-Ringot",email = "christophe.ringot1996@gmail.com"}
]
readme = "README.md"
requires-python = "3.13"
dependencies = [
    "requests (>=2.32.4,<3.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "numpy (>=2.3.1,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "tqdm (>=4.67.1,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
